requirements.txt
asgiref==3.5.2
configparser==5.3.0
Django==4.1.2
et-xmlfile==1.1.0
numpy==1.21.3
openpyxl==3.0.9
pandas==1.5.1
Pillow==8.4.0
PyMySQL==1.0.2
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.3
